\name{pTReCASE}
\alias{pTReCASE}
\alias{purTReCASE}
\docType{package}
\title{
pTReCASE R Package
}
\description{
This package is designed to implement the pTReCASE model fitting procedure, which models total read count (TReC)
and allele specific expression (ASE) in impure tumor tissues to identify tumor specific eQTL. Utilizing prior estimates of tumor purity, pTReCASE estimates tumor-specific eQTL effects using either a pTReC or pTReCASE model determined by use of a diagnostic Cis-Trans score test.
}
\details{
\tabular{ll}{
Package: \tab pTReCASE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-01-20\cr
License: \tab \cr
}
}
\author{
Douglas Roy Wilson, Jr.

Maintainer: Wei Sun, Ph.D. (wsun@fredhutch.org)
}

\keyword{ package }

